refactor(worker): relocate customer-fault classifier to neutral module (LFE-14990)
Move the classification core (classifyCustomerFault, isCustomerFaultError, the
code sets and cause-chain walk) from blobstorage/isCustomerFaultError.ts to a
neutral integrations/customerFaultClassification.ts so PostHog can share it.
isCustomerFaultError.ts becomes a thin re-export and keeps only the
blob-specific disable metric — a non-behavioral move (blob's runtime behavior
is unchanged; its existing test still passes).
